\name{tablaxlsx.estilos.set}
\alias{tablaxlsx.estilos.set}
\title{Write a matrix or data.frame with styles in a sheet of a excel workbook}
\description{
Puts the styles used by default in escribirTabla
}
\usage{
tablaxlsx.estilos.set(estilos.defecto)
}

\arguments{
  \item{estilos.defecto}{The list of styles to use by default in escribirTabla}

}
\details{
Puts the list estilos.defecto in the  tablaxlsx.Estilos option for use by default in escribirTabla. 
}

\seealso{\link{escribirTabla}, \link[openxlsx]{createStyle} and \link{estilosDGE}}

