\name{trim}
\alias{trim}
\title{It removes the spaces from the beginning or end of a text}
\description{
Returns a text or a vector of texts as the of the parameter but no spaces at the end or at the beginning of each text
}
\usage{
trim(x)
}

\arguments{
  \item{x}{Text o vector of texts}
}

\value{
text or a vector of texts as the of the parameter but no spaces at the end or at the beginning of each text
}

\examples{
## Example
texto=trim("  Esto es una prueba ")
}
