\name{verXL}
\alias{verXL}
\title{Show in excel a matrix or data.frame}
\description{
Show in excel a matrix or data.frame
}
\usage{
verXL(tabla,...)
}

\arguments{
  \item{tabla}{Matrix, data.frame or any object that can handle escribirTabla }
  \item{...}{parameters of escribirTabla } 
}

\examples{
## Example
\dontrun{
  verXL(iris,fila=1,columna=1)
}
}
